+++ coreweave/object_storage/resource_bucket_policy.go
@@ -85,7 +85,7 @@ func waitForBucketPolicy(parentCtx context.Context, client *s3.Client, bucket st
 		return err
 	}
 
-	return pollUntil("bucket versioning configuration", parentCtx, 5*time.Second, 5*time.Minute, func(ctx context.Context) (bool, error) {
+	return coreweave.PollUntil("bucket versioning configuration", parentCtx, 5*time.Second, 5*time.Minute, func(ctx context.Context) (bool, error) {
 		out, err := client.GetBucketPolicy(ctx, &s3.GetBucketPolicyInput{Bucket: aws.String(bucket)})
 		if err != nil {
 			return false, err
